+++ NEWS.md
@@ -6,6 +6,9 @@ variable and fixed positions within sample (Issue
 https://github.com/surh/HMVAR/issues/4)
 4. Added executable to subset SNPs from MIDAS after merging 
 (Issue https://github.com/surh/HMVAR/issues/3).
+6. read_midas_data now can keep non CDS positions.
+(Issue https://github.com/surh/HMVAR/issues/7)
+7. Created midas_to_bimbam function
 
 # 0.0-4
 1. Added ggtree import